split 7k and bluefin7k
